requirements-dev.txt
# Development dependencies for watchtower SDK

# Core dependencies
google-adk>=0.1.0

# Testing
pytest>=7.0
pytest-asyncio>=0.21
pytest-cov>=4.0

# Code quality
black>=23.0
mypy>=1.0
ruff>=0.1.0

# Type stubs
types-PyYAML

# Optional for config file support
PyYAML>=6.0

